Django Microservices Application/device_app/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==4.3.0
django-countries==7.5.1
django-crispy-forms==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
dsinternals==1.2.4
impacket==0.9.23
opencv-python==4.7.0.72
sqlparse==0.4.3
requests==2.26.0